fix prefix file detection picking up other package files
