+++ tests/test-recipes/split-packages/python_test_dep/conda_build_config.yaml
@@ -1,0 +1,3 @@
+python:
+  - 2.7
+  - 3.5
