+++ tests/test-recipes/split-packages/python_test_dep/test_thingie.py
@@ -1,0 +1,1 @@
+print('weeee')
